refactor:#99 스키마 파일명 변경
