+++ src/lib/hooks/use-profile-form.ts
@@ -1,6 +1,6 @@
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { editProfileSchema } from '@/lib/utils/validation/auth-validate';
+import { editProfileSchema } from '@/lib/utils/validation/auth-shema';
 import type { EditFormData } from '@/types/components/edit-profile-form';
 
 export const useProfileForm = (initNickname: string) => {
