+++ src/lib/hooks/use-signup-form.ts
@@ -1,6 +1,6 @@
 import { useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
-import { SignupFormData, signupSchema } from '@/lib/utils/validation/auth-validate';
+import { SignupFormData, signupSchema } from '@/lib/utils/validation/auth-shema';
 
 export const useSignupForm = () => {
   const form = useForm<SignupFormData>({
